requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1625643473416/work
certifi==2021.5.30
colorama==0.4.4
Django==3.0.3
django-debug-toolbar==3.2.4
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==9.8.2
halo==0.0.31
log-symbols==0.0.14
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
spinners==0.0.24
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
termcolor==1.1.0
text-unidecode==1.3
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
